pyproject.toml
[project]
name = "air-quiality-vs-asthma"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=6.29.5",
    "kaggle==1.6.17",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "polars>=1.29.0",
    "pycountry>=24.6.1",
    "pyyaml>=6.0.2",
    "tensorboard>=2.19.0",
    "torch==2.7.0",
    "torchinfo>=1.8.0",
    "torchvision==0.22.0",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu118", marker = "sys_platform == 'linux'" },
]
torchvision = [
  { index = "pytorch-cpu", marker = "sys_platform != 'linux'" },
  { index = "pytorch-cu118", marker = "sys_platform == 'linux'" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true
